refactor: cliにサブコマンドで処理を指示できるように改良
- cli をサブコマンドで複数の機能が実行できるように改良
- ソースコードを機能毎に spec と cmd に分離
